+++ src/components/attribution-editor.js
@@ -26,6 +26,9 @@ class AttributionEditor extends React.Component {
       }
     )
   }
+  componentWillReceiveProps (props) {
+    this.setState({block: props.initialBlock})
+  }
   render () {
     const {block} = this.state
     const {type, metadata} = block
@@ -42,7 +45,7 @@ class AttributionEditor extends React.Component {
         , { className: 'AttributionEditor-metadata'
           , style:
             { width: '90%'
-            , margin: '-48px auto 36px'
+            , margin: '0 auto 36px'
             , padding: '1.5em 3em 0'
             , background: '#fff'
             , border: '1px solid #ddd'
@@ -155,8 +158,6 @@ AttributionEditor.childContextTypes =
 AttributionEditor.propTypes =
   { initialBlock: React.PropTypes.object.isRequired
   , id: React.PropTypes.string.isRequired
-  , imgfloConfig: React.PropTypes.object
-  , store: React.PropTypes.object
   }
 export default React.createFactory(AttributionEditor)